+++ CHANGELOG.md
@@ -5,12 +5,14 @@ FEATURES
 * **New Analyzer:** `schemavalidatefunc`: returns `SchemaValidateFunc` declarations
 * **New Check:** `R005`: check for `ResourceData.HasChange()` calls that can be combined into one `HasChanges()` call
 * **New Check:** `S021`: check for `Schema` that should omit `ComputedWhen`
+* **New Check:** `S022`: check for `Schema` of `TypeMap` with invalid `Elem` of `*schema.Resource`
 * **New Check:** `V001`: check for custom `SchemaValidateFunc` that implement `validation.StringMatch()` or `validation.StringDoesNotMatch()`
 
 ENHANCEMENTS
 
 * cmd/tfproviderlint: Add `-V` and `-version` flags for version information (#40)
 * helper/astutils: Functions for determining package functions, package receiver methods, package types, and some function parameter types
+* helper/terraformtype: Support pointers with `IsHelperResourceTypeTestCase()`, `IsHelperResourceTypeTestStep()`, `IsHelperSchemaTypeResource()`, `IsHelperSchemaTypeResourceData()`, `IsHelperSchemaTypeSchema()`, and `IsHelperSchemaTypeSet()` functions
 * passes: Add `AllChecks` variable which can be used to bootstrap custom downstream linters
 
 # v0.6.0